Architecture of API Films.
> New Film Class.
> Route empty files. ( CRUD )
> WriteJson empty
> old index is now words
> New index.ts
