+++ src/modules/Film/Film.ts
@@ -1,0 +1,6 @@
+export class Film {
+
+  constructor(public name: string, public autor: string, public year: number) {}
+
+  
+}
